capstone/requirements.txt
Django>=3.2.8
psycopg2-binary>=2.8
gunicorn>=20.0.4
python-dotenv

